GPUTimelineManager and uniform management of frame delay (#2097)
* Added GPUTimelineManager (renderer-side equivalent to the DeferredDropper) for executing callbacks n frames later.
* Added callbacks to RenderScope so it manages calling begin/end frame for clients.

+++ crates/lgn-graphics-renderer/src/resources/persistent_descriptor_set_manager.rs
@@ -112,7 +112,7 @@ impl PersistentDescriptorSetManager {
     }
 
     pub fn frame_update(&mut self) {
-        self.render_frame = (self.render_frame + 1) % self.num_render_frames as u64;
+        self.render_frame = (self.render_frame + 1) % self.num_render_frames;
         self.bindless_index_allocator
             .release_indexes(&self.removed_indices[self.render_frame as usize]);
         self.removed_indices[self.render_frame as usize].clear();
